This adds some pom cleanups:
- Use vecmath 1.3.1
- Take JDK version from the parent pom

Also some code changes:
- Improvements to CrystalTransform class & xtal package
- Warnings and logging improvements

Conflicts:
	biojava3-alignment/pom.xml
	biojava3-core/pom.xml
	biojava3-modfinder/pom.xml
	biojava3-phylo/pom.xml
	biojava3-protein-disorder/pom.xml

+++ biojava3-alignment/src/main/java/org/biojava3/alignment/template/AbstractPairwiseSequenceAligner.java
@@ -150,7 +150,7 @@ public abstract class AbstractPairwiseSequenceAligner<S extends Sequence<C>, C e
     }
 
     @Override
-    protected short getSubstitutionScore(int queryColumn, int targetColumn) {
+    protected int getSubstitutionScore(int queryColumn, int targetColumn) {
         return getSubstitutionMatrix().getValue(query.getCompoundAt(queryColumn), target.getCompoundAt(targetColumn));
     }
 
@@ -173,8 +173,8 @@ public abstract class AbstractPairwiseSequenceAligner<S extends Sequence<C>, C e
             for (C c : target) {
                 maxt += getSubstitutionMatrix().getValue(c, c);
             }
-            max = (short) Math.max(maxq, maxt);
-            score = min = isLocal() ? 0 : (short) (2 * getGapPenalty().getOpenPenalty() + (query.getLength() +
+            max = (int) Math.max(maxq, maxt);
+            score = min = isLocal() ? 0 : (int) (2 * getGapPenalty().getOpenPenalty() + (query.getLength() +
                     target.getLength()) * getGapPenalty().getExtensionPenalty());
         }
     }
